More type fixes

+++ Math/Geometry/Shape/D3/Sphere.php
@@ -116,7 +116,7 @@ final class Sphere implements D3ShapeInterface
      */
     public static function getRadiusByVolume(float $v) : float
     {
-        return (float) pow($v * 3 / (4 * pi()), 1 / 3);
+        return pow($v * 3 / (4 * pi()), 1 / 3);
     }
 
     /**
